rcp: Add an rcp project to generate a launcher
The idea is to be able to generate a Trace Compass (Incubator) executable.
For example, it could be used to deploy a Trace Compass branded for Android
trace analysis. The file are copied from org.eclipse.tracecompass.rcp.
This is more a proof of concept right now since it introduces a lot of code duplication.

Change-Id: Ibbac04b6b4e3cd2e7e0017fbe0401bb98ad0f27a
Signed-off-by: Guillaume Champagne <guillaume.champagne@polymtl.ca>
Signed-off-by: Alexis-Maurer Fortin <alexis-maurer.fortin@polymtl.ca>
Signed-off-by: Hugo Genesse <hugo.genesse@polymtl.ca>
Signed-off-by: Pierre-Yves Lajoie<pierre-yves.lajoie@polymtl.ca>
Signed-off-by: Eva Terriault <eva.terriault@polymtl.ca>